DrinkIngrForm for DrinkEntry done and encapsulate DateTimeInput
